requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
feedparser
langdetect
chardet
pdfminer.six
tiktoken
datasketch
PyYAML
